% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.R
\docType{data}
\name{bd}
\alias{bd}
\title{Coordenadas de estabelecimentos da cadeia de SAN em Campos dos Goytacazes}
\format{
Um data.frame com 8353 linhas e 4 colunas:
\describe{
\item{lat}{numeric. Latitude}
\item{lng}{numeric. Longitude}
\item{nome}{character. Nome do estabelecimento}
\item{tipo}{character. Tipo do estabelacimento}
}
}
\usage{
bd
}
\description{
Um banco de dados contendo as coordenadas de latitude e longitude de 7.789
unidades agrícolas, 519 escolas públicas, 35 supermercados e 10 feiras livres
em Campos dos Goytacazes, RJ, Brasil
.
}
\keyword{datasets}
